backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-genai
Pillow
PyMuPDF
python-dotenv
pydantic
pytest
httpx
